added a sketch for immediate mode nodes. Those nodes could be operations and output that value instead of the subject to be more in line with the original immediate mode API. Context input and output pins are mainly in order to express execution order. Window Begin/End, Style push/pop brackets will be modeled as regions. An advantage of this approach is that you can avoid calling patches (inside those regions), which can hold parts of your business logic, not only the ImGui commands. So, when a certain part of the UI reports to be collapsed / invisble the whole region is disabled... This is very much in line with the original idea. Further advantages are that nodes like "GetContentRegionMax" can be called imperatively at the right moment and the documentation of their usage can be covered by the original library. Immediate mode patches can be read from top to bottom. Imperatively and can be compared directly with the textual code examples. In the retained mode it is rather unclear how to integrate a node like "GetContentRegionMax" into your scene graph.
If the immediate mode turns out to work fine: discuss behavior subject outputs, operation nodes, ref inputs of the original lib, possibly ref links as lang feature of VL, Skia immediate mode.
Another interesting approaches to test: only retained mode node set, as is, but with a region that allows the usage of nodes like "GetContentRegionMax" that now will always refer to the outer (parent) widget. The user only uses such a region when there is a need to clarify the context for nodes like "GetContentRegionMax"